refactor: presenter에서 데이터 생성자 주입받도록 구조 개선
