+++ app/src/main/java/woowacourse/movie/view/home/theater/TheaterContract.kt
@@ -2,7 +2,6 @@ package woowacourse.movie.view.home.theater
 
 import woowacourse.movie.domain.Movie
 import woowacourse.movie.domain.Showings
-import woowacourse.movie.domain.Theater
 
 interface TheaterContract {
     interface Presenter {
@@ -13,14 +12,5 @@ interface TheaterContract {
         fun handleInvalidTicket()
 
         fun showTheaterList(showings: List<Showings>)
-    }
-}
-
-class TheaterPresenter(
-    private val view: TheaterContract.View,
-) : TheaterContract.Presenter {
-    override fun fetchData(movie: Movie) {
-        val showings = movie.let { Theater.findTheatersShowingMovie(it.title) }
-        view.showTheaterList(showings)
     }
 }
